double to float

+++ src/edu/wpi/first/wpilibj/templates/commands/LeftRodDoNothing.java
@@ -24,7 +24,7 @@ public class LeftRodDoNothing extends CommandBase {
 
     // Called repeatedly when this Command is scheduled to run
     protected void execute() {
-        double s;
+        float s;
         s=leftrod.getRodServo();
         leftrod.setRodServoValue(s);
          SmartDashboard.putNumber("Doing nothing left Rod servo value=",s );
