+++ src/edu/wpi/first/wpilibj/templates/commands/RightRodDoNothing.java
@@ -24,7 +24,7 @@ public class RightRodDoNothing extends CommandBase {
 
     // Called repeatedly when this Command is scheduled to run
     protected void execute() {
-        double s;
+        float s;
         s=rightrod.getRodServo();
         rightrod.setRodServoValue(s);
          SmartDashboard.putNumber("Doing nothing right Rod servo value=",s );
